Corrections to MS portability changes that broke Mac builds: 1. computeInformationScores (in main/phylotesting.cpp), call to max    No longer compiled on Macintosh (updated it so it will). 2. PhyloTree::computeLikelihoodDervEigenSIMD, likewise 3. PhyloTree::computeLikelihoodBranchEigenSIMD, likewise 4. PhyloTree::computePartialParsimonySankoffSIMD, no need to free    tip_buffer (in Macintosh builds), that is only needed in MS builds.

+++ main/phylotesting.cpp
@@ -440,7 +440,7 @@ string detectSeqTypeName(string model_name) {
 
 void computeInformationScores(double tree_lh, int df, size_t sample_size, double &AIC, double &AICc, double &BIC) {
 	AIC = -2 * tree_lh + 2 * df;
-	AICc = AIC + 2.0 * df * (df + 1) / max(sample_size - df - 1, 1);
+	AICc = AIC + 2.0 * df * (df + 1) / max(static_cast<int>(sample_size - df - 1), 1);
 	BIC = -2 * tree_lh + df * log(sample_size);
 }
